+++ package.json
@@ -27,6 +27,7 @@
         "mongoose": "^5.11.9",
         "morgan": "^1.10.0",
         "passport-local": "^1.0.0",
-        "path": "^0.12.7"
+        "path": "^0.12.7",
+        "react-router-dom": "^5.2.0"
     }
 }
